run_servers_and_cl.py: find a non-existent model directory to save

+++ src/scripts/run_servers_and_cl.py
@@ -2,6 +2,7 @@
 
 import os
 import yaml
+import math
 import time
 import argparse
 import signal
@@ -31,21 +32,29 @@ def run_ttp(ttp_path, yaml_settings_path):
         old_model_dir = fingerprint['abr_config']['model_dir']
         model_parent_dir = path.dirname(old_model_dir)
 
-        new_model_base = cc + '-' + datetime.utcnow().strftime('%Y%m%d')
-        new_model_dir = path.join(model_parent_dir, new_model_base)
+        new_model_base_prefix = cc + '-' + datetime.utcnow().strftime('%Y%m%d')
+        # increment i until a non-existent directory is found
+        i = 0
+        while True:
+            i += 1
+            new_model_base = new_model_base_prefix + '-' + str(i)
+            new_model_dir = path.join(model_parent_dir, new_model_base)
+            if not path.isdir(new_model_dir):
+                break
 
         # run ttp.py
         start_time = time.time()
-        print('Continual learning: loaded {} and training {}'
-              .format(path.basename(old_model_dir), new_model_base))
+        sys.stderr.write('Continual learning: loaded {} and training {}\n'
+                         .format(path.basename(old_model_dir), new_model_base))
 
         check_call([ttp_path, yaml_settings_path, '--cl', '--cc', cc,
                     '--load-model', old_model_dir,
                     '--save-model', new_model_dir])
 
         end_time = time.time()
-        print('Continual learning: new model {} is available after {:.2f} hours'
-              .format(new_model_base, (end_time - start_time) / 3600))
+        sys.stderr.write(
+            'Continual learning: new model {} is available after {:.2f} hours\n'
+            .format(new_model_base, (end_time - start_time) / 3600))
 
         # update model_dir
         fingerprint['abr_config']['model_dir'] = new_model_dir
@@ -53,7 +62,7 @@ def run_ttp(ttp_path, yaml_settings_path):
     # write YAML settings with updated model_dir back to disk
     with open(yaml_settings_path, 'w') as fh:
         yaml.safe_dump(yaml_settings, fh, default_flow_style=False)
-    print('Updated model_dir in {}'.format(yaml_settings_path))
+    sys.stderr.write('Updated model_dir in {}\n'.format(yaml_settings_path))
 
 
 def main():
@@ -70,27 +79,35 @@ def main():
     # execute run_servers
     run_servers_proc = Popen([run_servers_path, yaml_settings_path],
                              preexec_fn=os.setsid, stderr=DEVNULL)
-    print('Started run_servers')
+    sys.stderr.write('Started run_servers\n')
 
-    while True:
-        # sleep until next CL_HOUR
-        td = datetime.utcnow()
-        wakeup = datetime(td.year, td.month, td.day, CL_HOUR, 0)
-        if wakeup <= td:
-            wakeup += timedelta(days=1)
+    try:
+        while True:
+            # sleep until next CL_HOUR
+            td = datetime.utcnow()
+            wakeup = datetime(td.year, td.month, td.day, CL_HOUR, 0)
+            if wakeup <= td:
+                wakeup += timedelta(days=1)
 
-        print('Sleeping until {} (UTC) to perform continual learning'
-              .format(wakeup))
-        time.sleep((wakeup - td).total_seconds())
+            sys.stderr.write('Sleeping until {} (UTC) to perform '
+                             'continual learning\n'.format(wakeup))
+            time.sleep(math.ceil((wakeup - td).total_seconds()))
 
-        # perform continual learning!
-        run_ttp(ttp_path, yaml_settings_path)
+            # perform continual learning!
+            run_ttp(ttp_path, yaml_settings_path)
 
-        # kill and restart run_servers with updated YAML settings
-        os.killpg(os.getpgid(run_servers_proc.pid), signal.SIGTERM)
-        run_servers_proc = Popen([run_servers_path, yaml_settings_path],
-                                 preexec_fn=os.setsid, stderr=DEVNULL)
-        print('Killed and restarted run_servers with updated YAML settings')
+            # kill and restart run_servers with updated YAML settings
+            if run_servers_proc:
+                os.killpg(os.getpgid(run_servers_proc.pid), signal.SIGTERM)
+            run_servers_proc = Popen([run_servers_path, yaml_settings_path],
+                                     preexec_fn=os.setsid, stderr=DEVNULL)
+            sys.stderr.write('Killed and restarted run_servers with updated '
+                             'YAML settings\n')
+    except Exception as e:
+        # clean up in case on exceptions
+        if run_servers_proc:
+            os.killpg(os.getpgid(run_servers_proc.pid), signal.SIGTERM)
+        print(e, file=sys.stderr)
 
 
 if __name__ == '__main__':
